Add load_vars action plugin (#19)
### Motivation
It would be nice to be able to use `vars_files` and `include_vars` with sops-encrypted YAML/JSON files. Unfortunately, that does not work, since there are no vault plugins yets (both directives/actions work with raw files and Ansible Vault, but not with anything else).

### Changes description
This adds a simple `load_vars` plugin, which behaves similarly to `include_vars`, but decodes sops-encrypted YAML/JSON files.

It only imports the file parts of `include_vars`, not the directory parts.

Also, by default, it does NOT treat the loaded values as variables, but as facts. This can be changed by setting `static: false`, but this functionality relies on a hack. I used another name (I originally wanted to use `include_sops_vars`) so the "proper" name can be used once this can be properly implemented without a hack.